I c constr fix

+++ inc/biot_mfedd.h
@@ -207,7 +207,7 @@ namespace dd_biot
 		unsigned long n_flux_mortar;
 		unsigned long n_pressure_mortar;
 		unsigned long n_Elast_mortar;
-		unsigned long n_mortar_dofs;
+		unsigned int n_mortar_dofs;
 
         // Subdomain coordinates (assuming logically rectangular blocks)
         Point <dim> p1;
